changed directory structure
i changed the directory structure to help with organization and compile
times. changing something in the ui, dsa, or game libraries won't cause
the whole project to recompile, just what was changed.

added a test function in window.h, defined it in window.cpp. the test
function uses SFML to create a window and draw a circle into it. then i
bring it into main.cpp to draw